<==== Mercurial data ====>
Branch: cloud_dev

+++ mediaserver_core/src/core/resource_management/manual_camera_searcher.cpp
@@ -442,8 +442,10 @@ QnManualCameraSearchProcessStatus QnManualCameraSearcher::status() const
         case QnManualResourceSearchStatus::CheckingOnline:
         {
             Q_ASSERT(m_hostRangeSize);
-            int currentProgress = m_hostRangeSize ?
-                m_ipChecker.hostsChecked() * PORT_SCAN_MAX_PROGRESS_PERCENT / m_hostRangeSize : 0;
+            int currentProgress = m_hostRangeSize
+                ? (int) m_ipChecker.hostsChecked() * PORT_SCAN_MAX_PROGRESS_PERCENT
+                    / m_hostRangeSize
+                : 0;
 
             result.status = QnManualResourceSearchStatus(m_state, currentProgress, MAX_PERCENT);
